fix: Delegate unhandled messages to the focused model
Currently, chatPage does not handle anything other than special events.
As a result, API/command responses from child models are passed to the
parent and processed without being communicated to the child (for
example, it is currently possible to open an external editor, but the
new text after closing the editor is not reflected in Crush's editor).
As an alternative solution, it is possible to export the child's tea.Msg
to the parent to address this issue, but we would prefer to handle the
logic closed in the child without communicating the details to the
parent as much as possible.